Доделал контроллеры

+++ src/main/java/com/example/person_service/controller/IndividualsController.java
@@ -3,8 +3,11 @@ package com.example.person_service.controller;
 import com.example.common.IndividualDto;
 import com.example.person_service.entity.Individual;
 import com.example.person_service.service.IndividualsService;
+import com.example.person_service.utils.UUIDValidator;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Mono;
 
 import java.util.UUID;
@@ -18,25 +21,32 @@ public class IndividualsController {
     private final IndividualsService individualsService;
 
     @PostMapping
-    public Mono<Individual> createIndividual (@RequestBody IndividualDto individualDto){
+    public Mono<Individual> createIndividual(@RequestBody IndividualDto individualDto) {
         return individualsService.createIndividual(individualDto);
     }
 
-    @PutMapping
-    public Mono<Individual> updateIndividual (@RequestBody IndividualDto individualDto){
-        return individualsService.updateIndividual(individualDto);
+    @PutMapping("/{id}")
+    public Mono<Individual> updateIndividual(@PathVariable String id, @RequestBody IndividualDto individualDto) {
+        if (UUIDValidator.isValidUUID(id)) {
+            return individualsService.updateIndividual(UUID.fromString(id), individualDto);
+        }
+        return Mono.error(new ResponseStatusException(HttpStatus.BAD_REQUEST));
     }
 
     @GetMapping("/{id}")
-    public Mono<Individual> getIndividualById (@PathVariable String id){
-        //todo: добавить валидацию на id, что это uuid
-        return individualsService.findById(UUID.fromString(id));
+    public Mono<Individual> getIndividualById(@PathVariable String id) {
+        if (UUIDValidator.isValidUUID(id)) {
+            return individualsService.findById(UUID.fromString(id));
+        }
+        return Mono.error(new ResponseStatusException(HttpStatus.BAD_REQUEST));
     }
 
     @DeleteMapping("/{id}")
-    public Mono<Individual> deleteIndividualById (@PathVariable String id){
-        //todo: добавить валидацию на id, что это uuid
-        return individualsService.deleteById(UUID.fromString(id));
+    public Mono<Individual> deleteIndividualById(@PathVariable String id) {
+        if (UUIDValidator.isValidUUID(id)) {
+            return individualsService.deleteById(UUID.fromString(id));
+        }
+        return Mono.error(new ResponseStatusException(HttpStatus.BAD_REQUEST));
     }
 
 }
